+++ src/renderer/RenderPass/Compute/ComputeSkinning/ComputeSkinningPass.h
@@ -16,7 +16,7 @@ namespace vg::renderer
         ComputeSkinningPass();
         ~ComputeSkinningPass();
 
-        void Setup(const gfx::RenderPassContext & _renderContext, double _dt) override;
+        void Setup(const gfx::RenderPassContext & _renderContext, float _dt) override;
         void Render(const gfx::RenderPassContext & _renderContext, gfx::CommandList * _cmdList) const override;
 
         void clearSkins();
